new zlib, better CMake integration, code fixups
Improvements:
 - zlib version 1.2.8 now part of the project
 - actually compiles when using a project generated by CMake for VS2013
 - CMake integration was almost completely rewritten, DLLs from VS now
   have names recognized by CascLib.h for automatic inclusion
 - fixed the few warnings that were generated
 - generated files are now put into a sane folder structure. Intermediate
   files, sadly, arent.

Potential Problems:
 - no more INSTALLing with CMake. I dont see a problem with reintroducing
   this at a later date, i just cant think of a good way at the moment.
 - support for Linux and Mac in CMake is lacking, mostly due to my
   infamiliarity with the OSs and GCC/CLang (or really any compiler other
   than VS). Someone with more knowledge about those platforms should give
   this a closer look, it probably doesnt work too well.

+++ src/CascBuildCfg.cpp
@@ -270,7 +270,7 @@ static int LoadInfoVariable(PQUERY_KEY pVarBlob, const char * szLineBegin, const
 
     // Initialize the blob
     pVarBlob->pbData = CASC_ALLOC(BYTE, (szLinePtr - szLineBegin) + 1);
-    pVarBlob->cbData = (size_t)(szLinePtr - szLineBegin);
+    pVarBlob->cbData = static_cast<DWORD>(szLinePtr - szLineBegin);
 
     // Check for success
     if(pVarBlob->pbData == NULL)
